+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/helpers/querybuilders/Expression.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders
+
+abstract class Expression {
+  abstract fun parameters(): List<String>
+}
