+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/helpers/querybuilders/conditions/Equal.kt
@@ -1,0 +1,16 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.conditions
+
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.Column
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.Condition
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.Expression
+
+class Equal<T>(private val column: Column<T>, private val value: Expression) : Condition() {
+  override fun addCondition(condition: Condition): Unit = throw IllegalStateException("Can't add a nested condition to the eq Operator")
+
+  override fun parameters(): List<String> = value.parameters()
+
+  override fun toString(): String = when (value) {
+    null -> "${column.name} is null" // TODO use value,
+    else -> "${column.name} = $value"
+  }
+}
