+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/helpers/querybuilders/conditions/Or.kt
@@ -1,0 +1,3 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.conditions
+
+class Or : CompositeCondition("OR")
