+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/helpers/querybuilders/functions/AthenaFunctions.kt
@@ -1,0 +1,14 @@
+package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.functions
+
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.expressions.Function
+import uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.helpers.querybuilders.expressions.Parameter
+import java.time.ZonedDateTime
+
+object AthenaFunctions {
+  fun fromIso8601Timestamp(timestamp: ZonedDateTime): Function = Function(
+    "from_iso8601_timestamp",
+    listOf(
+      Parameter(timestamp),
+    ),
+  )
+}
